extract opgenie client code to new file, rename to main
